Basic wrapper for the InjectMan
I was working on the InjectMan. I Added passing of parameters, added the
math for computation from cartesian space to motor positions and back.
The basic wrapper for the InjectMan is now finished. There are plenty
of TODO's in the code, many of those can be decided later if and when we
need such features.